Fix GIF Playback Issue by Restricting Placeholder Assignment
This commit addresses a performance issue in GIF playback. By introducing a condition `isReachMaxRepeatCount`, the assignment of `previousFrame` to a placeholder frame is now only performed when the maximum repeat count of the GIF is reached. This change prevents unnecessary placeholder assignments in infinite GIF animations, optimizing the playback performance.